feat: error refactoring
