+++ api/account/v2alpha1/account_sdk.go
@@ -180,7 +180,7 @@ func (r *ListSSHKeysResponse) UnsafeGetTotalCount() uint32 {
 
 // UnsafeAppend should not be used
 // Internal usage only
-func (r *ListSSHKeysResponse) UnsafeAppend(res interface{}) (uint32, scw.SdkError) {
+func (r *ListSSHKeysResponse) UnsafeAppend(res interface{}) (uint32, errors.SdkError) {
 	results, ok := res.(*ListSSHKeysResponse)
 	if !ok {
 		return 0, errors.New("%T type cannot be appended to type %T", res, r)
