+++ api/k8s/v1beta3/k8s_sdk.go
@@ -625,7 +625,7 @@ func (r *ListClustersResponse) UnsafeGetTotalCount() uint32 {
 
 // UnsafeAppend should not be used
 // Internal usage only
-func (r *ListClustersResponse) UnsafeAppend(res interface{}) (uint32, scw.SdkError) {
+func (r *ListClustersResponse) UnsafeAppend(res interface{}) (uint32, errors.SdkError) {
 	results, ok := res.(*ListClustersResponse)
 	if !ok {
 		return 0, errors.New("%T type cannot be appended to type %T", res, r)
@@ -1068,7 +1068,7 @@ func (r *ListPoolsResponse) UnsafeGetTotalCount() uint32 {
 
 // UnsafeAppend should not be used
 // Internal usage only
-func (r *ListPoolsResponse) UnsafeAppend(res interface{}) (uint32, scw.SdkError) {
+func (r *ListPoolsResponse) UnsafeAppend(res interface{}) (uint32, errors.SdkError) {
 	results, ok := res.(*ListPoolsResponse)
 	if !ok {
 		return 0, errors.New("%T type cannot be appended to type %T", res, r)
@@ -1404,7 +1404,7 @@ func (r *ListNodesResponse) UnsafeGetTotalCount() uint32 {
 
 // UnsafeAppend should not be used
 // Internal usage only
-func (r *ListNodesResponse) UnsafeAppend(res interface{}) (uint32, scw.SdkError) {
+func (r *ListNodesResponse) UnsafeAppend(res interface{}) (uint32, errors.SdkError) {
 	results, ok := res.(*ListNodesResponse)
 	if !ok {
 		return 0, errors.New("%T type cannot be appended to type %T", res, r)
