+++ api/test/v1/test_sdk.go
@@ -315,7 +315,7 @@ func (r *ListHumansResponse) UnsafeGetTotalCount() uint32 {
 
 // UnsafeAppend should not be used
 // Internal usage only
-func (r *ListHumansResponse) UnsafeAppend(res interface{}) (uint32, scw.SdkError) {
+func (r *ListHumansResponse) UnsafeAppend(res interface{}) (uint32, errors.SdkError) {
 	results, ok := res.(*ListHumansResponse)
 	if !ok {
 		return 0, errors.New("%T type cannot be appended to type %T", res, r)
