+++ go.mod
@@ -4,5 +4,6 @@ go 1.12
 
 require (
 	github.com/dnaeon/go-vcr v1.0.1
+	github.com/pkg/errors v0.8.1
 	gopkg.in/yaml.v2 v2.2.2
 )
